+++ packages/flutter_meedu/example/lib/controllers/counter_controller.dart
@@ -18,9 +18,9 @@ class CounterController extends SimpleNotifier {
   }
 
   @override
-  void onDispose() {
+  void dispose() {
     print("counter disposed");
-    super.onDispose();
+    super.dispose();
   }
 }