+++ packages/flutter_meedu/example/lib/pages/tabs_page/tabs/home/controller.dart
@@ -4,10 +4,10 @@ class HomeTabController extends SimpleNotifier {
   final counter = 0.obs;
 
   @override
-  void onDispose() {
+  void dispose() {
     counter.close();
     print("HomeTabController disposed");
-    super.onDispose();
+    super.dispose();
   }
 }