+++ packages/flutter_meedu/example/lib/pages/video/video_page_controller.dart
@@ -19,9 +19,9 @@ class VideoPageController extends SimpleNotifier {
   }
 
   @override
-  void onDispose() {
+  void dispose() {
     _controller.dispose();
     print("video disposed");
-    super.onDispose();
+    super.dispose();
   }
 }